--- src/app/view-employee/employee.model.ts
@@ -1,10 +1,0 @@
-export interface Employee {
-    employeeId: number,
-    managerId: number,
-    firstName: String,
-    lastName: String,
-    email: String,
-    userName: String,
-    password: String,
-    managerType: Boolean
-}
